Log NRP version including git commit (#909)
* adapters/mc: Log NRP version during startup

Currently NRP don't contain the OsConfig version, which makes to correlate
incoming logs with source code. Add a log line that prints the version info
during loading.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>

* version.h: Include short git hash in version information

We don't have tags for released NRP versions, so this is the next best way to
determine what git commit a release was produced from.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>

* adapters/mc: Add command used to print NRP version to README

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>

---------

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>