MpiServer: Fix missing pthread_join
The return value of pthread_create was being assigned to the same variable
that stored the pthread id - the outcome was that g_mpiServerWorker had the
value 0 (no error on create). This cause the pthread_join to be skipped.

pthread_t is meant to be an opaque type, so store the return from pthread_create
and compare that to decide whether to join.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>