Add clang static analysis worfklow (#992)
* workflows: Add static analyis workflow

Using clang's builtin static analyzer.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>

* devops: Add clang-tools to ubuntu-22.04 container

These are used by the static analysis workflow.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>

* workflows: Add temporary install step to static-analysis workflow

Can be removed once Ubuntu 22.04 container image is rebuilt.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>

* adapters/pnp: Add missing free's in main()

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>

* logging: Add missing null check in TrimLog

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>

* Add missing frees() throughout the codebase

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>

* platform: mmiPayload is not guaranteed to its size

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>

* module/test: Fix memory leaks

Detected by static analysis. One case is a missing free() in the cleanup
of LoadModule(). The other case is that ParseModuleInfo() only performs
cleanup if validation all the json fields are correct, and if they aren't
cleanup is skipped.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>

* modules/compliance: Disable static analysis for move constructor tests

The tests check that the move constructor is implemented correctly, so
disable the use-after-move static check.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>

* workflows/static-analysis: Exclude json libs from analysis

CommandRunner includes rapidjson which triggers a false postive, and I haven't
found a way to exclude a system path from scanning. CommandRunner is not in
active development or use, so it is safe to skip.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>

workflows/static-analysis: Another attempt to exempt rapidjson

* modules/test: Check for null pointers before calling close

Either of .close and .session may be null, in which case we should not
attempt to call close.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>

* common: Skip capture of return that is not used after comparison

Found by static analysis.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>

* workflows/static-analysis: Generate ProcedureMap.cpp before analysis

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>

* SshUtils: Remove dead stores

- textResultLength is not used outside the comparison. Remove.
- textPrefixLength is not used outside the comparison but it involves
  arithmetic and helps readability. Keep it.
- in CheckSshProtocol() - the first two branches had dead stores, that
  were unconditionally replaced by the status assignment in the third
  if condition. Combine the branches into a sequence of if-else and
  remove the store to status. The logic is the same, because
  CheckLineFoundNotCommentedOut also performs a check for file existance.

* SshUtils: Set status to ENOENT if file does not exist

This will never happen because the IsSshServerActive() function checks for the
existence of the configuration file as well.

* SshUtils: Free memory in case strlen check fails

Bit of a dilemma here. The Formatting will always result in a strlen() > 0
but the static checker is right to complain here. If the strlen() were to
be 0, the rest of the function would do nonsense. So add a free.

* PassUtils: Remove unused assignment to status

* OtherUtils: Remove unused assignment to status

* FileUtils: Remove unused assignment to result

* PackageUtils: Remove unused assignment to status

* DaemonUtils: Remove unused assignment to result

* DeviceInfoUtils: Remove unused assignment to status variable

* platform/Main: Fix static analysis issues

writeResult is unused. This is an attempt to fix the static analysis issues
while making sure that the result from write is not discarded (which triggers
a unused-result warning).

* platform/MpiServer: Remove unused assignments

Overwritten in every loop iteration.

* logging: Use MT-safe localtime_r()

Localtime returns a pointer to a static buffer, which is not safe in a MT
context. The MpiServer uses a thread and this can get flagged by the thread
sanitizer as a data-race. Replace the usage with localtime_r.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>

* MpiServer: Fix missing pthread_join

The return value of pthread_create was being assigned to the same variable
that stored the pthread id - the outcome was that g_mpiServerWorker had the
value 0 (no error on create). This cause the pthread_join to be skipped.

pthread_t is meant to be an opaque type, so store the return from pthread_create
and compare that to decide whether to join.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>

* tests: Replace ASSERT with EXPECT

The ASSERT gets flagged as a potential mem leak (which is not correct) but
EXPECT does not get flagged the same way so switch to EXPECT.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>

* adapters/pnp: Fix static analysis issue

writeResult is never read but write() requires that the result not be ignored.
Move the UNUSED line lower down so that the static analyzer approves.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>

* common/logging: Remove default branch in switch

To eliminate "Value stored to 'result' during its initialization is never read"
warning.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>

* compliance/tests: Rework CreateFile to prevent leak on assert failure

This is arguably a false positive but it's not too much work to rework so that
we do free before asserting.

* platform: Free ModuleInfo on validation failure

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>

* adapters/mc: Fix static analysis issues

- miResult assignment is a dead store
- uninitialized OsConfigResource_SetTargetResource instance may be passed to
  Destruct call
- same for OsConfig_GetTargetResource

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>

* devops: centos-8: Fix centos-8 base image

* common: Eliminate unused local and use stronger typing for LoggingLevel

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>

---------

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>