complianceengine: Remove accidentally committed lower case procedure files
EnsureInteractiveUsersDotFilesAccessIsConfigured.cpp,
EnsureInteractiveUsersHomeDirectoriesAreConfigured.cpp and the corresponding
schema files were accidentally committed in two copies, with the duplicates
starting with lower case names. We didn't notice this because we use globbing
in cmake and this only became visible on Windows. Windows is not case sensitive
on file names, and git flagged the issue because two of the duplicates had
small differences from review.

Reconcile the file differences and remove the lower case files. In the
procedures, I also fixed up an issue that was missed in code review: several
checks for !result.HasValue() were followed by an access to result.Value()
which would case a memory error as result is not checked. These have been
replaced with a return.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>