Add cmake and Result safeguards (#1060)
* Result: Throw if Value/Error tag is incorrect on access

We had a case where the following code made it through code review:
```
if (!result.HasValue())
{
    // do anything other than return
}
// invalid access
if (result.Value() == ...)
{
    ...
}
```
To prevent these from slipping through - add these checks and throw when they
happen. The exception will be caught in our entrypoints and logged.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>

* cmake: Switch from globbing to file lists in complianceengine

Cmake recommends against using globs for collecting source files: globs are
only re-executed when cmake is re-run, and cmake will only re-run if any of the
CMakeLists.txt are touched (or when called explicitly). This makes it easy to
accidentally miss new files. It also makes it harder to audit when new files
are added, which happened recently when two files with the same symbols but
filenames differing in casing only were added.

To prevent these issues in the future switch to explicit lists of source files.
The globbing was always meant as a temporary solution to avoid conflicts, but
now that we have a significant number of files in the lists, conflicts are
much less likely. Include two sanity checks in cmake:

- check if there are any extra files in the source dir that are not part of the source list
- check if all procedures have a corresponding schema file

The Cmake minimum version bump is needed to silence a warning when using if
IN_LIST.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>

* GenProcedureMap: Add sanity check for no-duplicate-procedures

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>