asb: Check all 4 grub.cfg path locations in Bootloader Password audit
A user running ASB on Red Hat 9.5 reported that the "Ensure all bootloaders
have password protection enabled" audit was failing for them, even though
/boot/grub2/grub.cfg contains:
```
if [ -f ${prefix}/user.cfg ]; then
  source ${prefix}/user.cfg
  if [ -n "${GRUB2_PASSWORD}" ]; then
    set superusers="root"
    export superusers
    password_pbkdf2 root ${GRUB2_PASSWORD}
  fi
fi
```
and user.cfg exists and has GRUB2_PASSWORD defined.

The immediate cause for the failure is that the audit function only checked 3
paths, missing /boot/grub2/grub.cfg. Switch to global constants for the paths
and cover all 4 values that we have constants for.

The checks are still not bullet proof but there's little more we can do without
actually evaluating the users grub.cfg. So this audit is little more than a
nudge for the users to do the right thing to accomplish compliance.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>